整合github-oauth2

+++ server/model/entity/user.go
@@ -1,11 +1,11 @@
 package entity
 
 type User struct {
-	Id        int     `gorm:"primary_key" json:"id"`
-	Username  string  `json:"username" form:"username" gorm:"unique"`
-	Password  string  `json:"password" form:"password"`
-	Nickname  string  `json:"nickname" form:"nickname"`
-	Email     string  `json:"email" form:"email"`
-	Phone     string  `json:"phone" form:"phone"`
-	GithubUID float64 `json:"github_uid" form:"github_uid"`
+	Id        int    `gorm:"primary_key" json:"id"`
+	Username  string `json:"username" form:"username" gorm:"unique"`
+	Password  string `json:"password" form:"password"`
+	Nickname  string `json:"nickname" form:"nickname"`
+	Email     string `json:"email" form:"email"`
+	Phone     string `json:"phone" form:"phone"`
+	GithubUID int    `json:"github_uid" form:"github_uid"`
 }
