+++ server/service/user.go
@@ -27,7 +27,7 @@ func (us *UserService) List(user entity.User) (users []entity.User, err error) {
 	return
 }
 
-func (us *UserService) GetByGithubId(id float64) (u entity.User, err error) {
+func (us *UserService) GetByGithubId(id int) (u entity.User, err error) {
 	err = utils.DB.Where("github_uid=?", id).First(&u).Error
 	return
 }
